Removing ManifestLookup
This is something already handled in core Symfony. And, with the latest
version of Encore, the values in entrypoints.json are the final paths:
we do not need to look up their values as keys in manifest.json to get
the final values (at one point, a draft version of entrypoints.json DID
work like this). The assets ARE still processed through the Asset system
so that you can process them through asset package configuration if you
want.